Fix comments from review

+++ firmware/Core/Src/adcs_drivers/adcs_types_to_json.c
@@ -1241,9 +1241,9 @@ uint8_t ADCS_wheel_currents_struct_TO_json(const ADCS_wheel_currents_struct_t *d
 
     const int16_t snprintf_ret = snprintf(json_output_str, json_output_str_size,
         "{\"wheel1_current_uA\":%lu,\"wheel2_current_uA\":%lu,\"wheel3_current_uA\":%lu}",
-        (unsigned long)data->wheel1_current_microamps,
-        (unsigned long)data->wheel2_current_microamps,
-        (unsigned long)data->wheel3_current_microamps);
+        (uint32_t)data->wheel1_current_microamps,
+        (uint32_t)data->wheel2_current_microamps,
+        (uint32_t)data->wheel3_current_microamps);
 
     if (snprintf_ret < 0) {
         return 2; // snprintf encoding error
@@ -1262,10 +1262,10 @@ uint8_t ADCS_cubesense_currents_struct_TO_json(const ADCS_cubesense_currents_str
 
     const int16_t snprintf_ret = snprintf(json_output_str, json_output_str_size,
         "{\"csense1_3v3_uA\":%lu,\"csense1_sram_uA\":%lu,\"csense2_3v3_uA\":%lu,\"csense2_sram_uA\":%lu}",
-        (unsigned long)data->cubesense1_3v3_current_microamps,
-        (unsigned long)data->cubesense1_sram_current_microamps,
-        (unsigned long)data->cubesense2_3v3_current_microamps,
-        (unsigned long)data->cubesense2_sram_current_microamps);
+        (uint32_t)data->cubesense1_3v3_current_microamps,
+        (uint32_t)data->cubesense1_sram_current_microamps,
+        (uint32_t)data->cubesense2_3v3_current_microamps,
+        (uint32_t)data->cubesense2_sram_current_microamps);
 
     if (snprintf_ret < 0) {
         return 2; // snprintf encoding error
@@ -1284,8 +1284,8 @@ uint8_t ADCS_misc_currents_struct_TO_json(const ADCS_misc_currents_struct_t *dat
 
     const int16_t snprintf_ret = snprintf(json_output_str, json_output_str_size,
         "{\"cubestar_current_uA\":%lu,\"torquer_current_uA\":%lu,\"cubestar_temp_mC\":%d}",
-        (unsigned long)data->cubestar_current_microamps,
-        (unsigned long)data->torquer_current_microamps,
+        (uint32_t)data->cubestar_current_microamps,
+        (uint32_t)data->torquer_current_microamps,
         (int)data->cubestar_mcu_temperature_mdeg_celsius);
 
     if (snprintf_ret < 0) {
